requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
Pillow==10.1.0
requests==2.31.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
jinja2==3.1.2
aiofiles==23.2.1
greenlet==3.0.2
psycopg2-binary==2.9.9
bcrypt==3.1.7
passlib[bcrypt]==1.7.4
